Split off base containers to separate module. Modified the various import statements to use relative paths when they're all within the lib/src directory.
