+++ lib/gps_history.dart
@@ -11,6 +11,7 @@
 library gps_history;
 
 export 'src/base.dart';
+export 'src/base_containers.dart';
 export 'src/time.dart';
 export 'src/gpc_efficient.dart';
 export 'src/gpc_list_based.dart';
